fix(vt-daemon): split terminals/tests tmux suites into tests/tmux/ to clear the 15-child fanout gate
The new tmux-server-reaper.test.ts pushed terminals/tests to 16 immediate
children, tripping the directory-fanout static gate (tier-0) and the
codebase-shape immediate-child check (tier-1-health). Group the seven
tmux-* suites under tests/tmux/, mirroring the source tmux/ module, and
deepen their parent-relative imports one level. No limit was raised.
