+++ packages/systems/vt-daemon/src/agent-runtime/terminals/tests/tmux/tmux-preflight.test.ts
@@ -6,7 +6,7 @@ import {
     formatMissingTmuxMessage,
     resetTmuxPreflightCache,
     type TmuxPreflightDeps,
-} from '../tmux/tmux-preflight.ts'
+} from '../../tmux/tmux-preflight.ts'
 
 type FakeChildOpts = {
     exitCode?: number | null
