+++ packages/systems/vt-daemon/src/agent-runtime/terminals/tests/tmux/tmux-session-manager.test.ts
@@ -13,8 +13,8 @@ import {
     listSessions,
     pipePaneToFile,
     sendKeys,
-} from '../tmux/tmux-session-manager.ts'
-import {shellQuote} from '../util/shellQuote.ts'
+} from '../../tmux/tmux-session-manager.ts'
+import {shellQuote} from '../../util/shellQuote.ts'
 
 const sessions: Set<string> = new Set<string>()
 const tempDirs: Set<string> = new Set<string>()
